move instrument_redis_cmd function
