+++ ddtrace/contrib/internal/aredis/patch.py
@@ -5,7 +5,7 @@ import aredis
 import wrapt
 
 from ddtrace import config
-from ddtrace._trace.utils_redis import _instrument_redis_cmd
+from ddtrace.contrib.internal.redis_utils import _instrument_redis_cmd
 from ddtrace.contrib.internal.redis_utils import _instrument_redis_execute_pipeline
 from ddtrace.contrib.internal.redis_utils import _run_redis_command_async
 from ddtrace.internal.schema import schematize_service_name
